labs/ra09-interface-description/requirements.txt
ncclient
requests
PyYAML
urllib3
